repackaged the app into powertool
